# src/api/swagger/ein-server.yml
openapi: '3.0.3'
info:
  title: 교육정보 알리미
  version: '1.0'
tags: 
  - name: file
    description: 파일 업로드 관련
  - name: blacklist
    description: 블랙리스트 관련
  - name: board
    description: 게시판 관련
  - name: user
    description: 회원관련
  - name: login/out
  - name: popup
    description: 광고 팝업 관련
  - name: push
    description: 푸쉬 알람 관련
paths:
  /file:
    post:
      tags: [file]
      description: 파일 업로드
      responses:
        '200':
          description: OK
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary

  /login:
    post:
      tags: [login/out]
      responses:
        '200':
          description: OK
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: myks790@gmail.com
                password:
                  type: string
                  example: 비밀번호

  /logout:
    get:
      tags: [login/out]
      responses:
        '200':
          description: OK

  /authUser:
    post:
      summary: 인증 회원 가입 (개발중, db comany_number 수정 필요)
      tags: [user]
      responses:
        '201':
          description: OK
        '400':
          description: 예외 발생
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: myks790@gmail.com
                password:
                  type: string
                  example: 1234
                name:
                  type: string
                  example: 강상훈
                nickname:
                  type: string
                  example: 닉네임
                phone:
                  type: string
                  example: 01012345678
                address:
                  type: string
                  example: 제주도 특별자치도 제주시 진남로 99길 10
                detail address:
                  type: string
                  example: 101호
                locaion:
                  type: integer
                  example: 17
                push agree:
                  type: string
                  example: yes
                  enum:
                    - yes
                    - no
                company:
                  type: string
                  example: 회~사
                company_number:
                  type: string
                  example: 06412345678
                position:
                  type: string
                  example: 대리
                website:
                  type: string
                  example: yes
                  enum:
                    - yes
                    - no


  /user:
    post:
      summary: 회원 가입
      tags: [user]
      responses:
        '201':
          description: OK
        '400':
          description: 이미 존재하는 email 등 예외 발생시
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: myks790@gmail.com
                password:
                  type: string
                  example: 1234
                name:
                  type: string
                  example: 강상훈
                nickname:
                  type: string
                  example: 닉네임
                phone:
                  type: string
                  example: 01012345678
                address:
                  type: string
                  example: 제주도 특별자치도 제주시 진남로 99길 10
                detail_address:
                  type: string
                  example: 101호
                location_id:
                  type: integer
                  example: 17
                push_agree:
                  type: string
                  example: yes
                  enum:
                    - yes
                    - no


  /user/{email}:
    get:
      summary: "유저정보를 반환함 (수정 필요 : pw 빼기 필요한 정보만 남기기 )"
      tags: [user]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    type: string
                    format: email
                    example: myks790@gmail.com
                  nickname:
                    type: string
                    example: 닉네임
                  name:
                    type: string
                    example: 강모씨
                  phone:
                    type: string
                    example: '01012345678'
                  address:
                    type: string
                    example: 제주도 특별자치도 제주시 진남로 99길 10
                  detail adress:
                    type: string
                    example: 101호
                  role:
                    type: string
                    enum:
                      - 일반
                      - 인증
                      - 관리자
                    example: 인증
                  push agree:
                    type: string
                    enum:
                      - yes
                      - no
                    example: yes
                  email check:
                    type: string
                    enum:
                      - yes
                      - no
                    example: no
                  authuser_id:
                    type: integer
                    example: 1
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: error message

    put:
      summary: 유저 정보를 수정함 (개발중) 
      tags: [user]
      responses:
        '200':
          description: OK
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: error message
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  example: 비밀번호
                nickname:
                  type: string
                  example: 닉네임
                phone:
                  type: string
                  example: 01012345678
                address:
                  type: string
                  example: 제주도 특별자치도 제주시 진남로 99길 10
                addressDetail:
                  type: string
                  example: 101호
                pushAgree:
                  type: boolean
                  example: true
    parameters:
      - name: email
        in: path
        required: true
        schema:
          type: string
        example: myks790@gmail.com
  
  /board/{boardId}:
    parameters:
      - name: boardId
        in: path
        required: true
        schema:
          type: integer
        example: 1
    get:
      summary: 게시글을 반환함
      tags: [board]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  title:
                    type: string
                    example: 여름 코딩 캠프
                  content:
                    type: string
                    example: "여름 14박 15일 코딩코딩코딩코딩딩딩딩 문의 : ~~~~"
                  writer:
                    type: string
                    example: 강상훈
                  location:
                    type: string
                    example: 제주특별자치도
                  start_date:
                    type: string
                    example: 2020-07-01T09:12:28Z
                  end_date:
                    type: string
                    example: 2020-08-31T07:12:28Z
        '404':
          description: 콘텐츠를 찾을 수 없음
    delete:
      summary: 게시글을 삭제함
      tags: [board]
      responses:
        '204':
          description: no content 삭제 성공
        '404':
          description: not found 삭제할 게시글을 찾을 수 없음
    put:
      summary: 게시글 수정함
      tags: [board]
      responses:
        '204':
          description: OK
        '404':
          description: 수정할 게시글을 찾을 수 없음
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: 여름 코딩 캠프
                start_date:
                  type: string
                  example: 2020-07-01T09:12:28Z
                end_date:
                  type: string
                  example: 2020-07-21T09:12:28Z
                content:
                  type: string
                  example: 9박 10일 여름 코딩 캠프~~~ 
                location_id:
                  type: integer
                  example: 17
                major_id:
                  type: integer
                  example: 1
                target_id:
                  type: integer
                  example: 1
  
  /board:
    get:
      summary: 게시글 리스트를 반환함
      tags: [board]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  contents:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        title:
                          type: string
                          example: 여름 코딩 캠프
                        writer:
                          type: string
                          example: 강상훈
                  pageNumber:
                    type: integer
                    example: 1
                  totalPages:
                    type: integer
                    example: 1
                  pageSize:
                    type: integer
                    example: 10
      parameters:
        - name: auth
          in: query
          required: true
          schema:
            type: string
            example: yes
            enum:
              - yes
              - no
          description: 인증 게시판, 자유 게시판 선택 플래그
          example: yes
        - name: page
          description: 'default: 1'
          in: query
          schema:
            type: integer
          example: 1
        - name: pageSize
          description: 'default: 10'
          in: query
          schema:
            type: integer
          example: 10
        - name: location_id
          in: query
          schema:
            type: integer
          example: 17
        - name: major_id
          in: query
          schema:
            type: integer
          description: "major id(1: '경영,회계,사무', 2: '취업,창업', 3: '금융,보험', ...)"
          example: "1"
        - name: target_id
          in: query
          schema:
            type: integer
          description: "target id(1: 성인, 2: 유아, 3: 청소년)"
          example: "1"
    post:
      summary: 게시글 생성 (db 필드 imageUrl 필요)
      tags: [board]
      responses:
        '201':
          description: OK
        '400':
          description: 생성시 필요한 필수값이 없을 경우 발생
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                  example: 1
                title:
                  type: string
                  example: 여름 코딩 캠프
                start_date:
                  type: string
                  example: 2020-07-01T09:12:28Z
                end_date:
                  type: string
                  example: 2020-07-21T09:12:28Z
                content:
                  type: string
                  example: 14박 15일 여름 코딩 캠프~~~ 
                location_id:
                  type: integer
                  example: 17
                major_id:
                  type: integer
                  example: 1
                target_id:
                  type: integer
                  example: 1
                auth:
                  type: string
                  example: yes
                  enum: 
                    - yes
                    - no
  
  /board/major:
    get:
      tags: [board]
      summary: 대분류를 반환함
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: 1
                    name:
                      type: string
                      example: 경영, 회계, 사무
                    code:
                      type: string
                      example: 001
  
  /board/target:
    get:
      tags: [board]
      summary: 대상을 반환함
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: 1
                    name:
                      type: string
                      example: 성인
                    code:
                      type: string
                      example: 001
  
  /board/location:
    get:
      tags: [board]
      summary: 시도 리스트를 반환함
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: 1
                    name:
                      type: string
                      example: 서울특별시
                    code:
                      type: string
                      example: 001

  /blacklist:
    get:
      tags: [blacklist]
      summary: 블랙리스트 목록을 반환함
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  contents:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        reporter:
                          type: string
                          format: email
                          example: myks790@gmail.com
                        reportedUser:
                          type: string
                          format: email
                          example: test@test.com
                        content:
                          type: string
                          example: 욕설등 부적절한 발언을 해서 신고합니다.
                        createdAt:
                          type: string
                          example: 2020-07-01T09:12:28Z
                        status:
                          type: string
                          enum:
                            - yes
                            - no
                            - waiting
                          example: waiting
                  pageNumber:
                    type: integer
                    example: 1
                  totalPages:
                    type: integer
                    example: 1
      parameters:
        - name: status
          description: 차단 여부(null, yes, no)
          in: query
          schema:
            type: string
            example: null
            enum:
              - null
              - yes
              - no

        - name: page
          description: 'default: 1'
          in: query
          schema:
            type: integer
          example: 1
        - name: pageSize
          description: 'default: 10'
          in: query
          schema:
            type: integer
          example: 10

    post:
      summary: 블랙리스트 추가
      tags: [blacklist]
      responses:
        '201':
          description: OK
        '400':
          description: 예외 발생
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reporter:
                  type: string
                  format: email
                  example: myks790@gmail.com
                reportedUser:
                  type: string
                  format: email
                  example: swlee292@naver.com
                content:
                  type: string
                  example: 욕설등 부적절한 발언을 해서 신고합니다.
                
  /blacklist/{blacklistId}:
    put:
      summary: 블랙리스트 status 수정
      tags: [blacklist]
      responses:
        '200':
          description: OK
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  description: yes, no 만 선택 가능(기본값은 null)
                  example: 'yes'
                  enum:
                    - yes
                    - no
      parameters:
        - name: blacklistId
          description: id
          required: true
          in: path
          schema:
            type: integer
          example: 1

  /popup:
    get:
      tags: [popup]
      summary: 홍보 팝업 조회 (개발중)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  image:
                    type: string
                    example: myks790.iptime.org/images/popup/asldfkjafs.png
                  active:
                    type: string
                    enum:
                      - yes
                      - no
                    example: yes

    post:
      summary: 광고 이미지 변경 (개발중)
      tags: [popup]
      responses:
        '200':
          description: OK
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                image:
                  type: string
                  example: myks790.iptime.org/images/popup/asldfkjafs.png
                
    put:
      summary: 팝업 활성화 설정 (개발중)
      tags: [popup]
      responses:
        '200':
          description: OK
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                active:
                  type: string
                  description: yes, no 만 선택 가능
                  example: 'yes'
                  enum:
                    - yes
                    - no

  /board/{boardId}/comment:
    parameters:
      - name: boardId
        description: 게시글 id
        in: path
        required: true
        schema:
          type: integer
        example: 1

    get:
      tags: [board]
      summary: 익명 댓글 목록을 반환함 (개발중  !!!누나 wirterFlag를 그냥 sql로 만들어줘요~ ㅎㅎ!!! )
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  contents:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        content:
                          type: string
                          example: 졸업자인데 참여 가능한가요?
                        wirterFlag:
                          type: boolean
                          example: 1
                          description: 게시글 작성자 여부 확인 플래그

    post:
      summary: 익명 댓글 추가
      tags: [board]
      responses:
        '201':
          description: OK
        '400':
          description: 존재하지 않은 게시글에 등록하는 경우등 예외 상황에서 발생
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: integer
                  example: 1
                content:
                  type: string
                  example: 졸업자인데 참여 가능한가요?
  
  /board/{boardId}/comment/{commentId}:     
    parameters:
      - name: boardId
        description: 게시글 id
        in: path
        required: true
        schema:
          type: integer
        example: 1
      - name: commentId
        description: comment id
        in: path
        required: true
        schema:
          type: integer
        example: 1       
    put:
      summary: 익명댓글 내용 수정
      tags: [board]
      responses:
        '204':
          description: OK
        '404':
          description: not found
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                  example: 조기졸업인데 참여 가능한가요?
              
    delete:
      summary: 익명댓글 내용 삭제 (개발중 sql 버그 수정 필요)
      tags: [board]
      responses:
        '204':
          description: OK
        '404':
          description: not found